// apps/front/src/lib/types/formState.d.ts
export type SignUpFormState = {
  data: {
    name?: string;
    email?: string;
    password?: string;
  };
  errors: {
    name?: string[];
    email?: string[];
    password?: string[];
  };
  message?: string;
};

export type CreateCommentFormState =
  | {
      data?: {
        content?: string;
        postId?: number;
      };
      errors?: {
        content?: string[];
      };
      message?: string;
      ok?: boolean;
      open?: boolean;
    }
  | undefined;

export type PostFormState = {
  data?: {
    postId?: number;
    title?: string;
    content?: string;
    thumbnail?: File | null;
    tags?: string[];
    published?: boolean;
    previousThumbnailUrl?: string;
  };
  errors?: {
    title?: string[];
    content?: string[];
    thumbnail?: string[];
    tags?: string[];
    isPublished?: string[];
  };
  message?: string;
  ok?: boolean;
};